Removed EnterTurn and fixed unit test cases.

+++ Testing/BattleTests/StaminaMonitorTests.cs
@@ -62,7 +62,7 @@ namespace Testing.BattleTests
             battleManager.ExitTurn += Raise.Event<BattleManager.OnTurn>(battleManager,
                                                                            Substitute.For<ICombatList>(),
                                                                            Substitute.For<ICombatList>(),
-                                                                           1);                        
+                                                                           (uint)1);                        
             staminaMonitor.Stamina.Should().Be(1);
             battleManager.DidNotReceiveWithAnyArgs().Remove(null, BattleManager.BattleSide.Attack, ReportState.Entering);
             
@@ -71,7 +71,7 @@ namespace Testing.BattleTests
             battleManager.ExitTurn += Raise.Event<BattleManager.OnTurn>(battleManager,
                                                                            Substitute.For<ICombatList>(),
                                                                            Substitute.For<ICombatList>(),
-                                                                           2);                        
+                                                                           (uint)2);                        
             battleManager.Received(1).Remove(combatGroup, BattleManager.BattleSide.Attack, ReportState.OutOfStamina);           
         }
